+++ tailwind.config.ts
@@ -14,5 +14,11 @@ const config: Config = {
     },
   },
   plugins: [],
+  safelist: [
+    { 
+      pattern: /(border|text|bg)-(red|orange|amber|yellow|lime|green|emerald|teal|cyan|sky|blue|indigo|violet|purple|fuchsia|pink|rose|slate|gray|zinc|neutral|stone)/,
+      variants: ['hover']
+     },
+  ],
 };
 export default config;