+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/), and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.3.0] - 2025-02-28
+
+### Added
+
+- Added new Form components, including Form, OutlineFieldset, OutlineFieldsetLegend, InputContainer, OutlineInput, FormButton, LoadingSpinner, and corresponding Storybook documentation.
+- Added FormContext, allowing all Form components to maintain consistent styling easier
+- Modified introduction.mdx, advising users of necessary TailwindCSS configuration
+- Modified README.md
+
 ## [1.2.0] - 2025-02-25
 
 ### Added